Implement Prometheus metrics that can be used as an assertion
The idea of this building block is that in Streaming Nft we must use externally collected metrics as an assertion. Load Runner metrics can be useful to do a sanity check, but the proper assertion must be done based on external metrics. PrometheusMetric is responsible for quering Prometheus API and returning metric's value to the test.