Reimplement penstock metric and fetching mechanism.
Prometheus metrics are values that needs to be fetched from Prometheus. Previous implementation didn't show that intention. It was a blocking API and the user didn't have a clue what Penstock is doing in background. Cats' IO type is the best to show that intention. It represents lazy side effect that must be explicitly executed.