Breaking change: Redesign relationship between LoadRunner and backends.
In original version LoadRunner takes backend as an implicit parameter of run method, but takes a message, throughput and duration as a explicit constructor parameter. It should be opposite. Backend is the most important parameter in LoadRunner, because it defines what kind of messages can be send. What is more having it an implicit parameters makes an api hard to explore.
This change makes a backend an explicit constructor parameter of LoadRunner, so that users are aware that backend is needed and when they use send method, they know what kind of messages can be send.