Simplify api (#26)
* Replace KafkaMessage case class with `createProducerRecord` function.

KafkaMessage case class was responsible for simplifying creation of ProducerRecords, objects accepted by Kafka backend. However there is no reason to have a case class for that. In this case we just need a function that takes topic, key, value and returns a ProducerRecord.

* Breaking change: Redesign relationship between LoadRunner and backends.

In original version LoadRunner takes backend as an implicit parameter of run method, but takes a message, throughput and duration as a explicit constructor parameter. It should be opposite. Backend is the most important parameter in LoadRunner, because it defines what kind of messages can be send. What is more having it an implicit parameters makes an api hard to explore.
This change makes a backend an explicit constructor parameter of LoadRunner, so that users are aware that backend is needed and when they use send method, they know what kind of messages can be send.